+++ Source/ChironEngine/Source/Modules/ModuleID3D12.cpp
@@ -5,7 +5,7 @@
 
 #include "ModuleWindow.h"
 
-#include "DataModels/CommandQueue/CommandQueue.h"
+#include "DataModels/DX12/CommandQueue/CommandQueue.h"
 
 ModuleID3D12::ModuleID3D12() : _currentBuffer(0), _vSync(true), _tearingSupported(false)
 {
@@ -100,7 +100,7 @@ void ModuleID3D12::ResizeBuffers(unsigned newWidth, unsigned newHeight)
     UpdateRenderTargetViews();
 
     // ------------- DEPTH-STENCIL ---------------------------
-
+ 
     CreateDepthStencil(newWidth, newHeight);
 }
 
@@ -344,10 +344,10 @@ void ModuleID3D12::CreateDepthStencil(unsigned width, unsigned height)
 {
     D3D12_CLEAR_VALUE clearValue = {};
     clearValue.Format = DXGI_FORMAT_D32_FLOAT;
-    clearValue.DepthStencil.Depth = 1.0f;
-    clearValue.DepthStencil.Stencil = 0;
+    clearValue.DepthStencil = { 1.0f, 0 };
     CD3DX12_HEAP_PROPERTIES heapProps = CD3DX12_HEAP_PROPERTIES(D3D12_HEAP_TYPE_DEFAULT);
-    CD3DX12_RESOURCE_DESC desc = CD3DX12_RESOURCE_DESC::Tex2D(DXGI_FORMAT_D32_FLOAT, width, height, 1, 0, 1, 0, D3D12_RESOURCE_FLAG_ALLOW_DEPTH_STENCIL);
+    CD3DX12_RESOURCE_DESC desc = CD3DX12_RESOURCE_DESC::Tex2D(DXGI_FORMAT_D32_FLOAT, width, height, 1, 0, 1, 0, 
+        D3D12_RESOURCE_FLAG_ALLOW_DEPTH_STENCIL);
 
     Chiron::Utils::ThrowIfFailed(
         _device->CreateCommittedResource(&heapProps, D3D12_HEAP_FLAG_NONE, &desc, D3D12_RESOURCE_STATE_DEPTH_WRITE,
@@ -364,6 +364,7 @@ void ModuleID3D12::CreateDepthStencil(unsigned width, unsigned height)
     D3D12_DEPTH_STENCIL_VIEW_DESC depthStencilDesc = {};
     depthStencilDesc.Format = DXGI_FORMAT_D32_FLOAT;
     depthStencilDesc.ViewDimension = D3D12_DSV_DIMENSION_TEXTURE2D;
+    depthStencilDesc.Texture2D.MipSlice = 0;
     depthStencilDesc.Flags = D3D12_DSV_FLAG_NONE;
 
     _device->CreateDepthStencilView(_depthStencilBuffer.Get(), &depthStencilDesc, dsvHandle);
